+++ 003_export_symbol/include/add_sub.h
@@ -1,5 +1,5 @@
 #ifndef _ADD_SUB_H_
 #define _ADD_SUB_H_
-long add_integer(long a, long b);
-long sub_integer(long a, long b);
+int add_integer(int a, int b);
+int sub_integer(int a, int b);
 #endif
